% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Adjust outcomes for differential trends}
\usage{
detrend(sdid, df)
}
\arguments{
\item{sdid}{sdid object containing the model to summarize.}

\item{df}{A data frame containing the variables in the model.}
}
\value{
data.frame
}
\description{
Adjust outcomes for differential trends
}
\examples{
# First fit a model using the original outcome, with or without covariate adjustments.
sdid_hosp <- sdid(hospitalized ~ cohort + yr + age + sex + comorb,
                  df = hosp,
                  intervention_var  = "intervention_yr")

# Calculate de-trending adjustments
hosp_det <- detrend(sdid = sdid_hosp,
                    df = hosp)

# Then refit the same model, substituting the _detrended version of the outcome
sdid_hosp_det <- sdid(hospitalized_detrended ~ cohort + yr + age + sex + comorb,
                      df = hosp_det,
                      intervention_var = "intervention_yr")
}
