% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_factor.R
\name{detrend_factor}
\alias{detrend_factor}
\title{Calculate aggregated variance to provide denominator for trend-adjusting outcomes}
\usage{
detrend_factor(sdid, df)
}
\arguments{
\item{sdid}{sdid object containing the model to summarize.}

\item{df}{A data frame containing the variables in the model.}
}
\value{
numeric
}
\description{
Calculate aggregated variance to provide denominator for trend-adjusting outcomes
}
\examples{
# First fit a model using the original outcome, with or without covariate adjustments.
sdid_hosp <- sdid(hospitalized ~ cohort + yr + age + sex + comorb,
                  df = hosp,
                  intervention_var  = "intervention_yr")

# Retrieve an estimate for the pre-intervention period
raw_pre_intervention_trend <-
  ave_coeff(sdid = sdid_hosp,
            coefs = select_period(sdid = sdid_hosp, period = "pre"))[["est"]]

# Normalize the pre-intervention estimate by de-trending factor
beta_detrend <-
  raw_pre_intervention_trend /
  detrend_factor(sdid = sdid_hosp,
                 df = hosp)
}
