% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have.R
\name{have}
\alias{have}
\title{Check whether a specified set of elements in a list exists}
\usage{
have(x, elements, all = TRUE)
}
\arguments{
\item{x}{list object}

\item{elements}{elements in x}
}
\value{
logical
}
\description{
Check whether a specified set of elements in a list exists
}
