% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_tsi.R
\name{id_tsi}
\alias{id_tsi}
\title{Identify time-since-intervention}
\usage{
id_tsi(df, cohort_var, time_var, intervention_var)
}
\arguments{
\item{df}{data frame containing the variables in the model.}

\item{cohort_var}{name of the variable in \code{df} that contains cohort assignments.}

\item{time_var}{name of the variable in \code{df} that contains time periods.}

\item{intervention_var}{name of the cohort-level variable in \code{df} that specifies which values in \code{time_var} correspond to the first post-intervention time period for each cohort.}
}
\description{
\code{id_tsi()} uses a supplied data set to identify the number of time periods away from the intervention each observation is. This is important for plotting and for aggregating model coefficients in \code{ave_coeff()}.
}
\examples{
# Generate a `tsi` object, containing a data frame showing the time since intervention
# (TSI value) for each time period in the data frame for each cohort.
id_tsi(hosp, cohort_var = "cohort", time_var = "yr", intervention_var = "intervention_yr")
}
