% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tsi.R
\name{new_tsi}
\alias{new_tsi}
\title{Constructor for a tsi object}
\usage{
new_tsi(df)
}
\arguments{
\item{df}{df containing times since intervention}
}
\value{
An object of class "tsi"
}
\description{
Constructor for a tsi object
}
