% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaggR.R
\name{optimizeStaggering}
\alias{optimizeStaggering}
\title{Optimize Staggering of Sample Preparations}
\usage{
optimizeStaggering(
  sampleNames = "auto",
  stepNames = "auto",
  timeToNextStep = c(),
  stepDuration = c(),
  nSamples = "auto",
  granularity = "auto",
  forceInterval = F,
  verbose = F
)
}
\arguments{
\item{sampleNames}{(Optional) Character vector of sample names; defaults to "auto" which creates
names like "Sample 1", "Sample 2", etc.}

\item{stepNames}{(Optional) Character vector of step names; defaults to "auto" which creates
names like "Step 1", "Step 2", etc.}

\item{timeToNextStep}{Numeric vector of the time between each step and the next. Must be 1 less than the length of stepNames.}

\item{stepDuration}{Numeric vector of durations (in minutes) required for each step.}

\item{nSamples}{Number of samples to process.}

\item{granularity}{(Optional) The time interval (in minutes) to test for optimal staggering;
defaults to "auto" "auto" to let the function decide.}

\item{forceInterval}{(Optional) If provided, forces the function to use a specific interval (in minutes).}

\item{verbose}{(Optional) Logical; if TRUE, prints progress details during computation.}
}
\value{
A list containing the optimal interval, several data frames with timecourse details,
  and a ggplot object of the scheduling.
}
\description{
This function computes the optimal time intervals for processing samples in a staggered
fashion, ensuring that overlapping steps are minimized.
}
\details{
This function simulates a time course schedule for processing multiple samples
  based on user-defined steps and durations. It checks for overlapping operations and
  incrementally adjusts the staggering interval until a valid schedule is found.
}
\examples{
\dontrun{
# Simple example with 3 steps and 4 samples:
res <- optimizeStaggering(
  stepNames = c("Start_treatment", "Load_samples_fuge", "Start_fuge", "Remove_supernatant"),
  timeToNextStep = c(59, 1, 15),
  stepDuration = c(15/60, 20/60, 30/60, 5),
  nSamples = 8,
  granularity = "auto",
  forceInterval = FALSE,
  verbose = TRUE
)
}
}
