% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_time_refs.R
\name{pick_time_refs}
\alias{pick_time_refs}
\title{Identify time period referents for cases when the user doesn't specify but instead names cohorts with suffixes corresponding to intervention time periods.}
\usage{
pick_time_refs(df, cohort_var, cohort_ref, time_var, time_offset = -1)
}
\arguments{
\item{df}{data frame containing the variables in the model.}

\item{cohort_var}{string specifying the name of the column in `df` that defines the intervention cohorts. If using this function, each cohort must be named to match the value of `time_var` that corresponds to the intervention period.}

\item{cohort_ref}{an optional string specifying the value of `cohort_var` to be used as the referent in the model. If not specified, the value is taken from the first observed value in `cohort_var`.}

\item{time_var}{string specifying the name of the column in `df` that defines time periods over the study.}

\item{time_offset}{integer specifying which time period relative to the intervention time period should be used as the referent for each cohort. Defaults to -1, which corresponds to the time period immediately preceding intervention.}
}
\value{
list
}
\description{
Identify time period referents for cases when the user doesn't specify but instead names cohorts with suffixes corresponding to intervention time periods.
}
