% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid.R
\name{sdid}
\alias{sdid}
\title{Fit a staggered difference-in-differences model}
\usage{
sdid(
  formula,
  df,
  weights = NULL,
  cohort_var = NULL,
  cohort_ref = NULL,
  cohort_time_refs = NULL,
  time_var = NULL,
  time_ref = NULL,
  intervention_var,
  .vcov = stats::vcov,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.}

\item{df}{data frame containing the variables in the model.}

\item{weights}{an optional vector of weights to be passed to `lm()` to be used in the fitting process. Should be NULL or a numeric vector.}

\item{cohort_var}{name of the variable in `df` that contains cohort assignments. If NULL, this is assumed to be the first column named in the right hand side of `formula`.}

\item{cohort_ref}{value of `cohort_var` that serves as the referent for main effects for cohorts. If NULL, this is assumed to the be the first value in the set of values for `cohort_var`.}

\item{cohort_time_refs}{a list, whose elements are named to match levels of `cohort_var`, specifying the value of `time_var` that serves as the referent for each time interaction with values of `cohort_var`. See 'Details.'}

\item{time_var}{name of the variable in `df` that contains time periods. If NULL, this is assumed to be the second column named in the right hand side of `formula`.}

\item{time_ref}{value of `time_var` that serves as the referent for main effects for time periods. If NULL, this is assumed to the be the first value in the set of values for `time_var`.}

\item{intervention_var}{name of the cohort-level variable in `df` that specifies which values in `time_var` correspond to the first post-intervention time period for each cohort.}

\item{.vcov}{Function to be used to estimate the variance-covariance matrix. Defaults to stats::vcov}

\item{...}{additional arguments to be passed to `.vcov`.}
}
\value{
sdid
}
\description{
Fit a staggered difference-in-differences model
}
