% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_period.R
\name{select_period}
\alias{select_period}
\title{Retrieve a list of interaction terms from a sdid model representing the pre-intervention period}
\usage{
select_period(sdid, period = "post", cohorts = NULL)
}
\arguments{
\item{sdid}{sdid object}

\item{period}{one of 'pre' or 'post', to return the pre-intervention or post-intervention coefficients respectively}

\item{cohorts}{character vector containing cohort levels to include in the term selection. If \code{cohorts} is omitted, all available cohorts will be selected}
}
\value{
character vector
}
\description{
Retrieve a list of interaction terms from a sdid model representing the pre-intervention period
}
