% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tsi.R
\name{select_tsi}
\alias{select_tsi}
\title{Retrieve a list of interaction terms from a sdid model representing a specified time-since-intervention}
\usage{
select_tsi(sdid, tsi = 0, cohorts = NULL)
}
\arguments{
\item{sdid}{A sdid object}

\item{tsi}{Integer indicating how many time periods away from the
intervention the function should identify coefficients for}

\item{cohorts}{A character vector containing cohort levels to include in the
term selection. If \code{cohorts} is omitted, all available cohorts will be
selected}
}
\value{
character vector
}
\description{
Retrieve a list of interaction terms from a sdid model representing a specified time-since-intervention
}
\examples{
# Fit a staggered difference-in-differences model
sdid_hosp <- sdid(hospitalized ~ cohort + yr + age + sex + comorb,
                  df = hosp,
                  intervention_var  = "intervention_yr")

# Select coefficients corresponding to 5 time periods before intervention
coef_selection_pre5 <- select_tsi(sdid_hosp,
                                  tsi = -5)
coef_selection_pre5

# Select coefficients corresponding to 2 time periods after intervention,
# for cohorts 5 and 6 only
coef_selection_post2 <- select_tsi(sdid_hosp,
                                   tsi = 2,
                                   cohorts = c("5", "6"))
coef_selection_post2
}
