% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggR-package.R
\docType{package}
\name{staggR-package}
\alias{staggR}
\alias{staggR-package}
\title{staggR: Fit Difference-in-Differences Models With Staggered Interventions}
\description{
A package for fitting linear difference-in-differences models in scenarios where intervention roll-outs are staggered over time. The package implements a version of an approach proposed by Sun and Abraham to estimate cohort- and time-since-treatment specific difference-in-differences parameters, and it provides convenience functions both for specifying the model and for flexibly aggregating coefficients to answer a variety of research questions.
}
\author{
\strong{Maintainer}: Kyle Hart \email{hartky@ohsu.edu}

Authors:
\itemize{
  \item Stephan Lindner
}

}
\keyword{internal}
