% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sdid.R
\name{summary.sdid_mdl}
\alias{summary.sdid_mdl}
\title{Summarize an sdid model}
\usage{
\method{summary}{sdid_mdl}(object, ...)
}
\arguments{
\item{object}{an \code{sdid_mdl} object.}

\item{...}{passed through.}
}
\value{
an object of class \code{summary.sdid_mdl}.
}
\description{
Summarize an sdid model
}
