% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_coeff.R
\name{ave_coeff}
\alias{ave_coeff}
\title{Aggregate a specified set of terms and corresponding standard errors from an sdid model object}
\usage{
ave_coeff(sdid, coefs)
}
\arguments{
\item{sdid}{sdid object containing the model to summarize}

\item{coefs}{character vector containing the names of coefficients to aggregate. Can be specified using \code{select_period()} or \code{select_terms()}.}
}
\value{
data.frame
}
\description{
Aggregate a specified set of terms and corresponding standard errors from an sdid model object
}
