% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_coeff.R
\name{ave_coeff}
\alias{ave_coeff}
\title{Prepare a data frame to work with sdid() function}
\usage{
ave_coeff(
  sdid,
  df,
  select_vars = NULL,
  selection = NULL,
  cohort_var,
  time_var,
  pop_var = NULL
)
}
\arguments{
\item{df}{data frame containing the variables in the model}

\item{select_vars}{list of variables selected for averaging; if NULL, then [SOMETHING] must be specified}

\item{pop_var}{name of variable showing population numbers (aggregate data only); if NULL, use nrow (i.e., number of rows = number of observations)}

\item{vcv}{variance-covariance corresponding to `mdl`}

\item{cohort_name}{character string for how cohorts are named. Must be specified if `select_vars` is NULL.}

\item{cohort_numbers}{list of cohort numbers selected for averaging. Must be specified if `select_vars` is NULL.}

\item{tsi_numbers}{list of time-since-intervention periods selected for averaging. Must be specified if `select_vars` is NULL.}
}
\value{
data.frame
}
\description{
Prepare a data frame to work with sdid() function
}
