% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_mdl.R
\name{new_sdid}
\alias{new_sdid}
\title{Constructor for a sdid_mdl object}
\usage{
new_sdid(mdl, vcov = NULL, tsi, key_vars)
}
\arguments{
\item{vcov}{A variance-covariance matrix}

\item{model}{A lm object}
}
\value{
An object of class "sdid_mdl"
}
\description{
Constructor for a sdid_mdl object
}
