% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_terms.R
\name{select_terms}
\alias{select_terms}
\title{Retrieve a list of interaction terms from a sdid model to be passed on for aggregation}
\usage{
select_terms(mdl, coefs = NULL, selection = NULL)
}
\arguments{
\item{mdl}{sdid object}

\item{coefs}{optional list of specific terms from \code{mdl} to be selected}

\item{selection}{list object containing values for named elements \code{cohorts}, \code{times}, and \code{tsi}. \code{cohorts} contains a character vector of cohort levels to include in the term selection; \code{times} contains a character vector of time period levels to include in the term selection; and \code{tsi} contains a vector of integers representing the number of units of time relative to each cohort's intervention to include in the term selection. If \code{cohorts} is omitted, all available cohorts will be selected. One of \code{times} or \code{tsi} must be specified. If both are specified, \code{times} is ignored.}
}
\value{
character vector
}
\description{
Retrieve a list of interaction terms from a sdid model to be passed on for aggregation
}
